Move native assets code into src/isolated/native_assets
